Refactor: layout 컴포넌트 헤더 유무로 구분
