+++ webapp/src/layouts/LayoutWithHeader.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import GlobalNavigation from 'components/GlobalNavigation';
 import { Outlet } from 'react-router-dom';
 import ToastNotificationProvider, {
   useToastNotificationAction,
@@ -7,11 +8,14 @@ import ToastNotificationProvider, {
 import WithProvider from 'hoc/withProvider';
 import { deleteMessage } from 'contexts/ToastNotification/action';
 import ToastNotification from 'components/ToastNotification';
-import * as S from './style';
+import * as S from './layout.style';
 
-export default WithProvider({ Providers: [ToastNotificationProvider], Component: Layout });
+export default WithProvider({
+  Providers: [ToastNotificationProvider],
+  Component: LayoutWithHeader,
+});
 
-function Layout() {
+function LayoutWithHeader() {
   const { toastList } = useToastNotificationState();
   const notifyDispatch = useToastNotificationAction();
   const deleteToastCallback = (id) => {
@@ -19,6 +23,9 @@ function Layout() {
   };
   return (
     <S.AppContainer>
+      <S.Header>
+        <GlobalNavigation />
+      </S.Header>
       <S.Main>
         <Outlet />
       </S.Main>
