+++ webapp/src/pages/Callback/ApiErrorCallback.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Button from 'components/Common/Button';
+import * as S from './callback.style';
 
 ApiErrorCallback.propTypes = {
   errorMessage: PropTypes.string.isRequired,
@@ -8,9 +10,11 @@ ApiErrorCallback.propTypes = {
 
 export default function ApiErrorCallback({ errorMessage, forceRefetch }) {
   return (
-    <div>
+    <S.Container>
       <h1>{errorMessage}</h1>
-      <button onClick={forceRefetch}>refetch</button>
-    </div>
+      <Button theme="primary" onClick={forceRefetch} customStyle={S.Button}>
+        재요청하기
+      </Button>
+    </S.Container>
   );
 }
