+++ webapp/src/pages/Callback/AuthErrorCallback.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useNavigate } from 'react-router-dom';
 import { ROUTE } from 'constant/route.constant';
+import Button from 'components/Common/Button';
+import * as S from './callback.style';
 
 AuthErrorCallback.propTypes = {
   errorMessage: PropTypes.string.isRequired,
@@ -14,9 +16,11 @@ export default function AuthErrorCallback({ errorMessage }) {
     naviagte(ROUTE.LOGIN);
   };
   return (
-    <div>
+    <S.Container>
       <h1>{errorMessage}</h1>
-      <button onClick={moveToLogin}>로그인 페이지로 이동하기</button>
-    </div>
+      <Button theme="primary" onClick={moveToLogin} customStyle={S.Button}>
+        재요청하기
+      </Button>
+    </S.Container>
   );
 }
