+++ webapp/src/router/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useRoutes } from 'react-router-dom';
-import Layout from 'layouts/layout';
+import LayoutWithHeader from 'layouts/LayoutWithHeader';
+import LayoutFullPage from 'layouts/LayoutFullPage';
 import publicRoutes from './PublicRoutes';
 import privateRoutes from './PrivateRoutes';
 import etcRoutes from './EtcRoutes';
@@ -8,10 +9,13 @@ import etcRoutes from './EtcRoutes';
 export default function Router() {
   const element = [
     {
-      element: <Layout />,
+      element: <LayoutWithHeader />,
       children: [...publicRoutes, ...privateRoutes],
     },
-    ...etcRoutes,
+    {
+      element: <LayoutFullPage />,
+      children: [...etcRoutes],
+    },
   ];
   return useRoutes(element);
 }
